chore: fixup - move document specific code to separate export
