+++ packages/gasket-nextjs/lib/gasket-data-provider.js
@@ -5,11 +5,8 @@ export const GasketDataContext = createContext({});
 
 /**
  * Provider for the GasketData, adds context to child elements.
- * @param {object} props - Props
- * @param {import('@gasket/data').GasketData} props.gasketData - Object of
- * GasketData
- * @param {ReactNode} props.children - Element to add GasketData context too
- * @returns {ReactElement} element
+ *
+ * @type {import('.').GasketDataProvider} GasketDataProvider
  */
 export const GasketDataProvider = ({ gasketData, children }) => {
   return createElement(GasketDataContext.Provider, { value: gasketData }, children);
