+++ packages/gasket-nextjs/lib/index.d.ts
@@ -1,29 +1,13 @@
 import type { GasketData } from '@gasket/data';
-import type { ComponentType, PropsWithChildren } from 'react';
-import type Document, {
-  DocumentContext,
-  DocumentInitialProps
-} from 'next/document';
-import type { IncomingMessage, ServerResponse } from 'http';
+import type { ComponentType, FunctionComponent, PropsWithChildren } from 'react';
 
 type SubstitutableHOC<T> = <C extends ComponentType<T>>(component: C) => C;
 
 /** Renders a script tag with JSON gasketData */
-export const GasketDataScript: ComponentType<{
+export const GasketDataScript: FunctionComponent<{
   /** Gasket data from response */
   data: GasketData;
 }>;
-
-/**
- * Make a wrapper to extend the Next.js Document, injecting a script with the
- * `gasketData` from the response object.
- */
-export function withGasketData(options?: {
-  /** Force script injection at particular index */
-  index?: number;
-}): (DocumentClass: typeof Document) => typeof DocumentClass & {
-  getInitialProps: typeof GasketDocumentGetInitialProps;
-};
 
 /** React hook that fetches GasketData in elements context and returns it */
 export function useGasketData(): GasketData;
@@ -31,24 +15,8 @@ export function useGasketData(): GasketData;
 /**
  * Provider for the GasketData, adds context to child elements.
  */
-export const GasketDataProvider: ComponentType<
+export const GasketDataProvider: FunctionComponent<
   PropsWithChildren<{ gasketData: GasketData }>
->;
-
-export declare type GasketDocumentContext = DocumentContext & {
-  res: ServerResponse<IncomingMessage> & {
-    locals: {
-      gasketData: GasketData;
-    };
-  };
-};
-
-export function GasketDocumentGetInitialProps(
-  ctx: GasketDocumentContext
-): Promise<
-  DocumentInitialProps & {
-    gasketData: GasketData;
-  }
 >;
 
 /**
@@ -57,3 +25,4 @@ export function GasketDocumentGetInitialProps(
  * NextDocument component.
  */
 export function withGasketDataProvider(): SubstitutableHOC<{}>;
+
