+++ packages/gasket-nextjs/test/helpers.js
@@ -6,7 +6,6 @@ const consoleError = console.error;
 /**
  * Silence noisy act warnings from console.error
  * TODO: Remove when fix is determined
- *
  * @returns {MockInstance<Console, any>} mock console.error
  */
 export function mockConsoleError() {
